Double precision in linear fitting module

+++ math/fittingfunctions/fityax.c
@@ -5,8 +5,8 @@
  * Fitting a linear function without intercept
  *
  * Parameters
- * xs: (float) vector of x s
- * ys: (float) vector of y s
+ * xs: (double) vector of x s
+ * ys: (double) vector of y s
  *
  * Return
  * fitting coefficients, a
@@ -29,8 +29,8 @@ fitresultyax* fityax(vector *xs, vector *ys)
   double x, y, xy = 0, x2 = 0;
 
   for(i = 0; i < xs->logLength; i++){
-    x = *(float*)vector_get(xs, i);
-    y = *(float*)vector_get(ys, i);
+    x = *(double*)vector_get(xs, i);
+    y = *(double*)vector_get(ys, i);
 
     xy += x * y;
     x2 += x * x;
